refactor: 클린 아키텍처 적용 - 클린 아키텍처 구조 적용 - 특정 변수명 camel case 로 수정
