+++ app/src/main/java/com/survivalcoding/stopwatch/domain/model/LapTimeRecord.kt
@@ -1,11 +1,11 @@
-package com.survivalcoding.stopwatch.database
+package com.survivalcoding.stopwatch.domain.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class LaptimeRecord(
+data class LapTimeRecord(
     @PrimaryKey(autoGenerate = true) var rid: Int = 0,
     @ColumnInfo(name = "elapsed_time") val elapsedTime: Int,
     @ColumnInfo(name = "end_time") val endTime: Int,
