+++ app/src/main/java/com/survivalcoding/stopwatch/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.stopwatch
+package com.survivalcoding.stopwatch.presentation
 
 import android.os.Bundle
 import androidx.activity.viewModels
@@ -9,6 +9,7 @@ import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.survivalcoding.stopwatch.R
 import com.survivalcoding.stopwatch.databinding.ActivityMainBinding
