+++ app/src/main/java/com/survivalcoding/stopwatch/presentation/fragment/StopWatchFragment.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.stopwatch.fragment
+package com.survivalcoding.stopwatch.presentation.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,10 +10,10 @@ import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.survivalcoding.stopwatch.MainViewModel
 import com.survivalcoding.stopwatch.R
-import com.survivalcoding.stopwatch.adapter.LaptimeRecordAdapter
 import com.survivalcoding.stopwatch.databinding.FragmentStopWatchBinding
+import com.survivalcoding.stopwatch.presentation.MainViewModel
+import com.survivalcoding.stopwatch.presentation.adapter.LaptimeRecordAdapter
 import java.text.DecimalFormat
 
 class StopWatchFragment : Fragment() {
