+++ app/src/main/java/com/survivalcoding/stopwatch/presentation/fragment/TimerFragment.kt
@@ -1,10 +1,10 @@
-package com.survivalcoding.stopwatch.fragment
+package com.survivalcoding.stopwatch.presentation.fragment
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.survivalcoding.stopwatch.R
 
 // TODO: Rename parameter arguments, choose names that match
