+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,27 +7,27 @@
 
     <fragment
         android:id="@+id/alarm"
-        android:name="com.survivalcoding.stopwatch.fragment.AlarmFragment"
+        android:name="com.survivalcoding.stopwatch.presentation.fragment.AlarmFragment"
         android:label="Alarm"
         tools:layout="@layout/fragment_alarm" />
     <fragment
         android:id="@+id/clock"
-        android:name="com.survivalcoding.stopwatch.fragment.ClockFragment"
+        android:name="com.survivalcoding.stopwatch.presentation.fragment.ClockFragment"
         android:label="Clock"
         tools:layout="@layout/fragment_clock" />
     <fragment
         android:id="@+id/timer"
-        android:name="com.survivalcoding.stopwatch.fragment.TimerFragment"
+        android:name="com.survivalcoding.stopwatch.presentation.fragment.TimerFragment"
         android:label="Timer"
         tools:layout="@layout/fragment_timer" />
     <fragment
         android:id="@+id/stopwatch"
-        android:name="com.survivalcoding.stopwatch.fragment.StopWatchFragment"
+        android:name="com.survivalcoding.stopwatch.presentation.fragment.StopWatchFragment"
         android:label="Stopwatch"
         tools:layout="@layout/fragment_stop_watch" />
     <fragment
         android:id="@+id/bedtime"
-        android:name="com.survivalcoding.stopwatch.fragment.BedtimeFragment"
+        android:name="com.survivalcoding.stopwatch.presentation.fragment.BedtimeFragment"
         android:label="Bedtime"
         tools:layout="@layout/fragment_bedtime" />
 </navigation>
